getting started again!
questions.md is brainstorming what we might do with this data.

test.py is a sanity check. it's looking at how to access the json data. it also found a bug in the data, which will probably require going back to the converter: 1999 season, first game, the data fails to show that one team missed the extra point after a touchdown.